Add files via upload
last version of the extended BDG script. Includes all recent changes across the team + corridor logic from the faculty. Data extraction was reformated to work with a one time recording process for all data. Some filters are still to be updated/ check. Python csv writer is also pending, though a simple csv creator is available as a panel preview at the end of the script.